+++ src/Components/Header/Header.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import './Header.css'
+import '../styles/Header.css'
 
 const Header = () => (
   <div className="top-header">
