+++ src/Components/Navbar/Navbar.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { Link } from 'react-router-dom'
-import './Navbar.css'
+import '../styles/Navbar.css'
 import logo from '../../assets/logo2.svg'
 
 const Navbar = () => {
